requirements.txt
streamlit
pandas
plotly
crewai
langchain-google-genai
google-generativeai
pillow